feat(profile): Extract language section 📝
